data type of Bild was changed to double

+++ src/main/java/Zahlenerkennung/Model/Bild.java
@@ -3,10 +3,10 @@ package Zahlenerkennung.Model;
 public class Bild {
 
     private final String label;
-    private final byte[] pixel;
+    private final double[] pixel;
     private final int rowOffset;
 
-    public Bild(String label, byte[] pixel, int rowOffset) {
+    public Bild(String label, double[] pixel, int rowOffset) {
         this.label = label;
         this.pixel = pixel;
         this.rowOffset = rowOffset;
@@ -16,7 +16,7 @@ public class Bild {
         return label;
     }
 
-    public byte[] getPixel() {
+    public double[] getPixel() {
         return pixel;
     }
